show all cards function
